requirements.txt
transformers==4.51.3
torch==2.7.0
pillow
datasets==3.6.0
tokenizers==0.21.1